:rewind: Revert ":recycle: MessquerschnittAnzahlInfo.vue in MessstelleInfo.vue integriert"
This reverts commit 0a25c7b7
