+++ frontend/src/components/messstelle/MessstelleInfo.vue
@@ -55,30 +55,6 @@
                 </span>
             </v-sheet>
         </v-row>
-
-        <v-row class="ma-0">
-            <v-sheet
-                v-if="!doMesssquerschnitteExist"
-                id="empty"
-                class="d-flex align-center justify-center px-4 pt-1"
-                color="transparent"
-            >
-                <h4 class="text-caption font-weight-bold">
-                    Für diese Messstelle sind keine weiteren Messquerschnitte
-                    vorhanden.
-                </h4>
-            </v-sheet>
-            <v-sheet
-                v-else
-                color="transparent"
-                class="px-4 py-2"
-            >
-                <h4 v-if="props.messstelle.messquerschnitte.length > 1">
-                    Informationen zu Messquerschnitten
-                </h4>
-                <h4 v-else>Informationen zum Messquerschnitt</h4>
-            </v-sheet>
-        </v-row>
     </v-container>
 </template>
 
@@ -127,10 +103,6 @@ const detektierteVerkehrsart = computed(() => {
 
 const abbauDatumExists = computed(() => {
     return props.messstelle.abbaudatum;
-});
-
-const doMesssquerschnitteExist = computed(() => {
-    return props.messstelle.messquerschnitte.length > 0;
 });
 
 const props = defineProps<Props>();
