+++ frontend/src/views/MessstelleView.vue
@@ -24,6 +24,7 @@
                     />
                     <MessstelleInfo :messstelle="messstelle" />
                     <OptionsmenueMessstelle :messstelle-id="messstelleId" />
+                    <MessquerschnittAnzahlInfo :messstelle="messstelle" />
                     <MessquerschnittInfo :messstelle="messstelle" />
                 </v-sheet>
             </v-col>
@@ -54,6 +55,7 @@ import MessquerschnittInfo from "@/components/messstelle/MessquerschnittInfo.vue
 import DefaultObjectCreator from "@/util/DefaultObjectCreator";
 import { ApiError } from "@/api/error";
 import { useStore } from "@/api/util/useStore";
+import MessquerschnittAnzahlInfo from "@/components/messstelle/MessquerschnittAnzahlInfo.vue";
 import OptionsmenueMessstelle from "@/components/messstelle/OptionsmenueMessstelle.vue";
 
 const messstelle: Ref<MessstelleInfoDTO> = ref(
